Generator will produce the output file on its own
We avoid using stdio redirection, because that pattern tends to
confuse make in the event that there is an error.  The output file is
created and so it looks like something was touched recently, even
though it is not the proper output file.  Having the generator select
the output on its own should avoid these failure modes.