Specify who signs the email in Generator
Previously, Generator signed an email with the key of the value of the
top-level From: header. However, we might want to sign an email with a
different key (to indicate that the top-level header has been tampered
with; see corpus/H.eml for example).

This commit allows callers to specify the identity of the signer when
calling Generator.sign(). See corpus/H.eml for an example use.